Fix code review stuff
