--- src/Warehouse/Application/Configuration/AddFluentValidatorConfigExtensions.cs
@@ -1,13 +1,0 @@
-﻿using FluentValidation;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace Application.Configuration
-{
-    public static class AddFluentValidatorConfigExtensions
-    {
-        public static void AddFluentValidatorConfiguration(this IServiceCollection services)
-        {
-            services.AddValidatorsFromAssembly(typeof(AddFluentValidatorConfigExtensions).Assembly);
-        }
-    }
-}
