--- src/Warehouse/Application/Exceptions/ModelValidationException.cs
@@ -1,15 +1,0 @@
-﻿using FluentValidation.Results;
-
-namespace Application.Exceptions
-{
-    public class ModelValidationException : Exception
-    {
-        public List<ValidationFailure> Errors { get; set; }
-
-        public ModelValidationException(List<ValidationFailure> errors)
-            : base(string.Join(',', errors.Select(f => f.ErrorMessage)))
-        {
-            Errors = errors;
-        }
-    }
-}
