+++ src/Warehouse/Application/Features/Commands/SyncProducts/SyncProductsCommand.cs
@@ -1,0 +1,5 @@
+﻿using MediatR;
+
+namespace Application.Features.Commands.SyncProducts;
+
+public record SyncProductsCommand(string sourceUri) : IRequest;
