--- src/Warehouse/Application/Features/Products/Query/GetProductsQuery.cs
@@ -1,10 +1,0 @@
-﻿using MediatR;
-
-namespace Application.Features.Products.Query
-{
-    public record GetProductsQuery(
-        double? MinPrice,
-        double? MaxPrice,
-        List<string> Highlights,
-        string Size) : IRequest<GetProductsResponse>;
-}
