--- src/Warehouse/Application/Features/Products/Synchronize/SyncProductsCommand.cs
@@ -1,6 +1,0 @@
-﻿using MediatR;
-
-namespace Application.Features.Products.Synchronize
-{
-    public record SyncProductsCommand(string sourceUri) : IRequest;
-}
