+++ src/Warehouse/Application/PipelineBehaviors/ValidationBehavior.cs
@@ -1,34 +1,33 @@
-﻿using Application.Exceptions;
+﻿using Application.Common.Exceptions;
 using FluentValidation;
 using MediatR;
 
-namespace Application.PipelineBehaviors
+namespace Application.PipelineBehaviors;
+
+public class ValidationBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
+    where TRequest : IRequest<TResponse>
 {
-    public class ValidationBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
-        where TRequest : IRequest<TResponse>
+    private readonly IEnumerable<IValidator<TRequest>> _validators;
+
+    public ValidationBehavior(IEnumerable<IValidator<TRequest>> validators)
     {
-        private readonly IEnumerable<IValidator<TRequest>> _validators;
+        _validators = validators;
+    }
 
-        public ValidationBehavior(IEnumerable<IValidator<TRequest>> validators)
+    public Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next, CancellationToken cancellationToken)
+    {
+        var context = new ValidationContext<TRequest>(request);
+        var failures = _validators
+            .Select(v => v.Validate(context))
+            .SelectMany(v => v.Errors)
+            .Where(v => v.Equals != null)
+            .ToList();
+
+        if (failures.Any())
         {
-            _validators = validators;
+            throw new ModelValidationException(failures);
         }
 
-        public Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next, CancellationToken cancellationToken)
-        {
-            var context = new ValidationContext<TRequest>(request);
-            var failures = _validators
-                .Select(v => v.Validate(context))
-                .SelectMany(v => v.Errors)
-                .Where(v => v.Equals != null)
-                .ToList();
-
-            if (failures.Any())
-            {
-                throw new ModelValidationException(failures);
-            }
-
-            return next();
-        }
+        return next();
     }
 }
