+++ src/Warehouse/Infrastructure/Cache/ICacheProvider.cs
@@ -1,9 +1,8 @@
-﻿namespace Infrastructure.Cache
-{
-    public interface ICacheProvider
-    {
-        T Get<T>(string cacheKey);
+﻿namespace Infrastructure.Cache;
 
-        void Set<T>(string cacheKey, T entity);
-    }
+public interface ICacheProvider
+{
+    T Get<T>(string cacheKey);
+
+    void Set<T>(string cacheKey, T entity);
 }
