+++ src/Warehouse/Infrastructure/Cache/InMemoryCacheProvider.cs
@@ -1,24 +1,23 @@
 ﻿using Microsoft.Extensions.Caching.Memory;
 
-namespace Infrastructure.Cache
+namespace Infrastructure.Cache;
+
+public class InMemoryCacheProvider : ICacheProvider
 {
-    public class InMemoryCacheProvider : ICacheProvider
+    private readonly IMemoryCache _memoryCache;
+
+    public InMemoryCacheProvider(IMemoryCache memoryCache)
     {
-        private readonly IMemoryCache _memoryCache;
+        _memoryCache = memoryCache;
+    }
 
-        public InMemoryCacheProvider(IMemoryCache memoryCache)
-        {
-            _memoryCache = memoryCache;
-        }
+    public T Get<T>(string cacheKey)
+    {
+        return (T)(_memoryCache.Get(cacheKey));
+    }
 
-        public T Get<T>(string cacheKey)
-        {
-            return (T)(_memoryCache.Get(cacheKey));
-        }
-
-        public void Set<T>(string cacheKey, T entity)
-        {
-            _memoryCache.Set(cacheKey, entity);
-        }
+    public void Set<T>(string cacheKey, T entity)
+    {
+        _memoryCache.Set(cacheKey, entity);
     }
 }
