--- src/Warehouse/Infrastructure/Http/Configuration/AddHttpConfigExtensions.cs
@@ -1,14 +1,0 @@
-﻿using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace Infrastructure.Http.Configuration
-{
-    public static class AddHttpConfigExtensions
-    {
-        public static void AddHttpConfiguration(this IServiceCollection services)
-        {
-            services.AddHttpClient();
-            services.AddTransient<IRetrieveProductsService, RetrieveProductsService>();
-        }
-    }
-}
