--- src/Warehouse/WebApi/Configuration/MemoryCacheExtensions.cs
@@ -1,14 +1,0 @@
-﻿using Infrastructure.Cache;
-
-namespace WebApi.Configuration
-{
-    public static class MemoryCacheExtensions
-    {
-        public static void AddMemoryCacheConfiguration(this IServiceCollection services)
-        {
-            services.AddMemoryCache();
-
-            services.AddSingleton<ICacheProvider, InMemoryCacheProvider>();
-        }
-    }
-}
