--- src/Warehouse/WebApi/DTOs/ProductsSummaryDto.cs
@@ -1,13 +1,0 @@
-﻿namespace WebApi.DTOs
-{
-    public class ProductsSummaryDto
-    {
-        public double MinPrice { get; set; }
-
-        public double MaxPrice { get; set; }
-
-        public List<string> Sizes { get; set; }
-
-        public List<string> Highlights { get; set; }
-    }
-}
