+++ src/Warehouse/WebApi/ErrorHandling/ExceptionHandlerExtensions.cs
@@ -1,36 +1,35 @@
-﻿using Application.Exceptions;
+﻿using Application.Common.Exceptions;
 using Microsoft.AspNetCore.Diagnostics;
 using static System.Net.Mime.MediaTypeNames;
 
-namespace WebApi.ErrorHandling
+namespace WebApi.ErrorHandling;
+
+public static class ExceptionHandlerExtensions
 {
-    public static class ExceptionHandlerExtensions
+    public static void ConfigureExceptionHandler(this IApplicationBuilder application)
     {
-        public static void ConfigureExceptionHandler(this IApplicationBuilder application)
+        application.Run(async context =>
         {
-            application.Run(async context =>
+            var exceptionHandlerPathFeature =
+                context.Features.Get<IExceptionHandlerPathFeature>();
+
+            if (exceptionHandlerPathFeature?.Error is ModelValidationException error)
             {
-                var exceptionHandlerPathFeature =
-                    context.Features.Get<IExceptionHandlerPathFeature>();
+                context.Response.StatusCode = StatusCodes.Status400BadRequest;
+                await context.Response.WriteAsJsonAsync(error.Errors
+                    .Select(e => new
+                    {
+                        e.PropertyName,
+                        e.ErrorMessage
+                    }));
+            }
+            else
+            {
+                context.Response.StatusCode = StatusCodes.Status500InternalServerError;
 
-                if (exceptionHandlerPathFeature?.Error is ModelValidationException error)
-                {
-                    context.Response.StatusCode = StatusCodes.Status400BadRequest;
-                    await context.Response.WriteAsJsonAsync(error.Errors
-                        .Select(e => new
-                        {
-                            e.PropertyName,
-                            e.ErrorMessage
-                        }));
-                }
-                else
-                {
-                    context.Response.StatusCode = StatusCodes.Status500InternalServerError;
-
-                    context.Response.ContentType = Text.Plain;
-                    await context.Response.WriteAsync("General exception was thrown.");
-                }
-            });
-        }
+                context.Response.ContentType = Text.Plain;
+                await context.Response.WriteAsync("General exception was thrown.");
+            }
+        });
     }
 }
