+++ src/Warehouse/WebApi/Models/Responses/QueryProductsResponse.cs
@@ -1,0 +1,8 @@
+﻿namespace WebApi.Models.Responses;
+
+public class QueryProductsResponse
+{
+    public ProductsSummaryResponse Summary { get; set; }
+
+    public List<ProductResponse> Products { get; set; }
+}
