+++ src/Warehouse/WebApi/Program.cs
@@ -1,23 +1,15 @@
-using Application.Configuration;
-using Infrastructure.Http.Configuration;
-using Infrastructure.Logger.Configurtion;
-using WebApi.Configuration;
+using Application.Extensions;
+using Infrastructure.Extensions;
 using WebApi.ErrorHandling;
+using WebApi.Extensions;
 
 var builder = WebApplication.CreateBuilder(args);
 
 builder.Logging.AddLogger(builder.Configuration);
 
-builder.Services.AddMemoryCacheConfiguration();
-builder.Services.AddQuartzConfiguration(builder.Configuration);
-builder.Services.AddHttpConfiguration();
-builder.Services.AddControllers();
-builder.Services.AddMediatrConfiguration();
-builder.Services.AddAutoMapper(typeof(Program));
-builder.Services.AddFluentValidatorConfiguration();
-
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
+builder.Services.AddApplication(builder.Configuration);
+builder.Services.AddInfrastructure();
+builder.Services.AddPresentation();
 
 var app = builder.Build();
 //Add Global exception handler
